content/base/test/test_bug675121.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=675121
-->
<head>
  <title>Test for Bug 675121</title>
  <script type="application/javascript" src="/MochiKit/packed.js"></script>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=675121">Mozilla Bug 675121</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 675121 **/
var callbackFired = false;
var xhrInProgress = false;
function f() {
  callbackFired = true;
  if (!xhrInProgress) {
    SimpleTest.finish();
  }
}

window.mozRequestAnimationFrame(f);
var xhr = new XMLHttpRequest();
xhr.open("GET", "file_bug675121.sjs", false);
xhrInProgress = true;
xhr.send();
xhrInProgress = false;
is(xhr.responseText, "Responded", "Should have a response by now");
is(callbackFired, false, "Callback should not fire during sync XHR");

if (!callbackFired) {
  SimpleTest.waitForExplicitFinish();
}
</script>
</pre>
</body>
</html>
